added authentication, added CRUD

+++ controllers/jobs_controller.py
@@ -1,7 +1,9 @@
 from flask import Blueprint, jsonify, request, abort
 from main import db
-from models import Job, Venue
-from schemas import jobs_schema, job_schema
+from models import Job, Venue, Application
+from schemas import jobs_schema, job_schema, application_schema
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from .helper_functions import authenticate_role_only
 
 jobs = Blueprint("jobs", __name__, url_prefix="/jobs")
 
@@ -26,10 +28,17 @@ def get_all_available_jobs_per_venue(venue_id):
 
 # post new jobs
 @jobs.route("/<int:venue_id>", methods=["POST"])
+@jwt_required()
+@authenticate_role_only(role="manager")
 def new_job(venue_id):
     venue = Venue.query.get(venue_id)
     if not venue:
         return abort(400, description="Venue does not exist")
+
+    # Check if manager has permission to post a job for this venue
+    if not str(venue.manager_id) == get_jwt_identity().replace("manager", ""):
+        return abort(401, description="You don't have permission to post for this venue.")
+
     job_fields = job_schema.load(request.json)
     job = Job(
         description=job_fields["description"],
@@ -46,11 +55,18 @@ def new_job(venue_id):
 
 
 # Delete a job
-@jobs.route("/<int:job_id>", methods=["DELETE"])
+@jobs.route("/delete/<int:job_id>", methods=["DELETE"])
+@jwt_required()
+@authenticate_role_only(role="manager")
 def delete_job(job_id):
     job = Job.query.get(job_id)
     if not job:
         return abort(400, description="Job does not exist")
+
+    # Return error if manager id does not match
+    job_venue = Venue.query.get(job.venue_id)
+    if not str(job_venue.manager_id) == get_jwt_identity().replace("manager", ""):
+        return abort(401, description="You don't have permission to delete this job.")
 
     db.session.delete(job)
     db.session.commit()
@@ -60,10 +76,19 @@ def delete_job(job_id):
 
 # Update a job
 @jobs.route("<int:job_id>", methods=["PUT"])
+@jwt_required()
+@authenticate_role_only(role="manager")
 def update_job(job_id):
     job = Job.query.get(job_id)
+    # Check if job exist in the database
     if not job:
         return abort(400, description="Job does not exist")
+
+    # Check if the job is managed by the manager
+    job_venue = Venue.query.get(job.venue_id)
+
+    if not str(job_venue.manager_id) == get_jwt_identity().replace("manager", ""):
+        return abort(401, description="You don't have permission to update this job.")
 
     job_fields = job_schema.load(request.json)
     job.description = job_fields.get("description", job.description)
@@ -71,6 +96,60 @@ def update_job(job_id):
     job.start_time = job_fields.get("start_time", job.start_time)
     job.finish_time = job_fields.get("finish_time", job.finish_time)
     job.pay_rate = job_fields.get("pay_rate", job.pay_rate)
+
+    db.session.commit()
+    return jsonify(job_schema.dump(job))
+
+
+# Apply for job
+@jobs.route("<int:job_id>/apply", methods=["POST"])
+@jwt_required()
+@authenticate_role_only(role="user")
+def apply_job(job_id):
+    job = Job.query.get(job_id)
+    if not job or job.status == "fulfilled":
+        return abort(400, description="The job does not exist or has been fulfilled.")
+
+    application = Application(
+        job_id=job_id,
+        barista_id=get_jwt_identity().replace("user", "")
+    )
+
+    db.session.add(application)
+    db.session.commit()
+    return jsonify(application_schema.dump(application))
+
+
+# Approve applicant
+@jobs.route("<int:job_id>/approve", methods=["PUT"])
+@jwt_required()
+@authenticate_role_only(role="manager")
+def approve_applicant(job_id):
+    # Check if job exist
+    job = Job.query.get(job_id)
+    if not job:
+        return abort(400, description="Job does not exist.")
+    # Check if the job is managed by the manager
+    job_venue = Venue.query.get(job.venue_id)
+    if not str(job_venue.manager_id) == get_jwt_identity().replace("manager", ""):
+        return abort(401, description="You don't have permission to update this job.")
+
+    # Check if there are applicants
+    applications = Application.query.filter_by(job_id=job_id).all()
+    if not len(applications):
+        return jsonify({"ERROR": "No applications were found."})
+
+    barista_id = application_schema.load(request.json)["barista_id"]
+    # Modify application status
+    for application in applications:
+        if application.barista_id == barista_id:
+            application.status = "approved"
+        else:
+            application.status = "unsuccessful"
+
+    # Change job status to fulfilled, also assign barista_id to job
+    job.status = "fulfilled"
+    job.barista_id = barista_id
 
     db.session.commit()
     return jsonify(job_schema.dump(job))
